requirements.txt
pytorch>=0.4.1
numpy>=1.14
matplotlib>=2.2
tqdm>=4.26.0
tick>=0.5.0
git+git://github.com/ManifoldFR/point-process-rust
